requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
selenium==4.15.2
fake-useragent==1.4.0
openai==1.3.7
lxml==4.9.3
numpy==1.24.3
python-dotenv==1.0.0
webdriver-manager==4.0.1